+++ tests/testthat/test-utilities-scenario-configuration.R
@@ -4,11 +4,12 @@ projectConfiguration <- createDefaultProjectConfiguration(path = "../data/Projec
 
 test_that("it throws an error when wrong scenario is defined", {
   scenarioNames <- "wrong"
-  expect_error(readScenarioConfigurationFromExcel(
-    scenarioNames = scenarioNames,
-    projectConfiguration = projectConfiguration
-  ),
-  regexp = messages$scenarioConfigurationNameNotFoundWhenReading(scenarioNames[[1]])
+  expect_error(
+    readScenarioConfigurationFromExcel(
+      scenarioNames = scenarioNames,
+      projectConfiguration = projectConfiguration
+    ),
+    regexp = messages$scenarioConfigurationNameNotFoundWhenReading(scenarioNames[[1]])
   )
 })
 
@@ -82,5 +83,68 @@ test_that("It creates multiple correct scenarios", {
   ))
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$simulationRunOptions, NULL)
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$simulationType, "Individual")
-  expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$steadyStateTime, 1000)
+  expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$steadyStateTime, 500)
+})
+
+test_that("It creates all scenarios if no name is defined", {
+  scenarioConfigurations <- readScenarioConfigurationFromExcel(
+    projectConfiguration = projectConfiguration
+  )
+  expect_equal(names(scenarioConfigurations), c("TestScenario", "TestScenario2"))
+})
+
+test_that("It correctly applies a custom function", {
+  scenarioNames <- "TestScenario"
+
+  scenarioConfigurations <- readScenarioConfigurationFromExcel(
+    scenarioNames = scenarioNames,
+    projectConfiguration = projectConfiguration
+  )
+
+  doseParamPath <- "Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose"
+  doseFactor <- 2
+  customFunction <- function(doseFactor, doseParamPath) {
+    doseParam <- getParameter(
+      path = doseParamPath,
+      container = simulation
+    )
+    doseParam$value <- doseParam$value * doseFactor
+  }
+  # Simulation without the custom function
+  sim <- initializeScenario(scenarioConfigurations[[1]])
+  # get the value of dose parameter
+  oldVal <- getQuantityValuesByPath(doseParamPath, sim)
+
+  scenarioConfigurations$TestScenario$customFunction <- customFunction
+  scenarioConfigurations$TestScenario$customFunctionArgs <- list(
+    doseFactor = doseFactor,
+    doseParamPath = doseParamPath
+  )
+
+  sim <- initializeScenario(scenarioConfigurations[[1]])
+
+  expect_equal(getQuantityValuesByPath(doseParamPath, sim), oldVal * doseFactor)
+})
+
+test_that("It throws an error when trying to set wrong arguments for the custom function", {
+  scenarioNames <- "TestScenario"
+
+  scenarioConfigurations <- readScenarioConfigurationFromExcel(
+    scenarioNames = scenarioNames,
+    projectConfiguration = projectConfiguration
+  )
+
+
+  customFunction <- function(doseFactor) {
+    doseParamPath <- "Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose"
+    doseParam <- getParameter(
+      path = doseParamPath,
+      container = simulation
+    )
+    doseParam$value <- doseParam$value * doseFactor
+  }
+
+  scenarioConfigurations$TestScenario$customFunction <- customFunction
+
+  expect_error(scenarioConfigurations$TestScenario$customFunctionArgs <- list(wrong = 2), messages$errorWrongArguments("doseFactor"))
 })